Migrate encoder tests to public Python APIs
